(fix): Add notifications (SignalR) and live-session features

+++ services/live-session-service/LiveSessionService.Application/Features/Common/AzuraCast/Models/AzuraCastCurrentSongData.cs
@@ -4,9 +4,9 @@ public sealed class AzuraCastCurrentSongData
 {
     public long ShId { get; init; }
     public AzuraCastSongData? Song { get; init; }
-    public long PlayedAt { get; init; }
-    public long Duration { get; init; }
-    public long Elapsed { get; init; }
-    public long Remaining { get; init; }
+    public double PlayedAt { get; init; }
+    public double Duration { get; init; }
+    public double Elapsed { get; init; }
+    public double Remaining { get; init; }
     public bool IsRequest { get; init; }
 }
